Fixes #21
The original fix in #21 was broken because the code that added the module to the cache would find "mscorlib" under it's simple name and assume that was the right one even when the fullname was different.  This now explicitly checks fullname first and only checks short name if fullname was not passed.  

Additionally the logic to determine which was the higher version meant that it never found duplicates if the 2nd version was lower than the 1st.  Now it always finds the highest, regardless of order.

Finally The relink has to happen after MonoModWasHere is added as MonoModWasHere also would reference the version of MonoMod.