Begin repository reorganization (stage 1)
This is the first of a few steps. Next, I will be adding a tools/ directory and moving
all of the MSBuild customization into there, and use Directory.Build.* to include them.
Following that, I will be completely reorganizing the MSBuild tooling to be more reusable,
less messy, and more complete.