Rename MonoMod.Common project to just Common
This now acts as a shared source location, and the project itself is
a NoTargets project. NoTargets projects are also automated now, with
tools/Common.NoTargets.props handling most of what such a project
would want automatically.

Common also now holds UnsafeAlias, and Common.CS.* automatically
import it along with a reference to ILHelpers, assuming it isn't
disabled in the target project.