Implement masked matching in BytePattern
It has one caveat though: there must be at least one non-masked byte in the pattern, or the
pattern will be treated as an empty pattern by the collection, and thus likely not ever match.
This is because BytePatternCollection currently does not/can not index masked patterns,
and so it doesn't search for the start of the pattern unless it has an unmasked byte.