Add debug assertions for ABI to PlatformTriple
This, in debug builds, automatically checks for discrepancies between the detected ABI and the
known ABI. Such a discrepancy could be due to BOTH a selftest being wrong AND the known
ABI being wrong. If one of these asserts fails, more testing is required to determine which is the
problem.