Fix potential out-of-bounds access in PlatformTriple.GetNativeMethodBodyWalk
We do this by asking the OS how much space we have which is actually readable. When we create
a Span for that region, we then know exactly how much memory is safe to access, guaranteeing
that our scan is safe.

Currently, the Span's length is capped at collection.MaxMinLength to prevent scanning into
other method thunks, but it may be prudent to instead scan for some arch-specific padding
pattern that is inserted between methods.