Fix DynamicData.For memory leak
When using DynamicData.For, the constructed DynamicData objects have the `keepAlive` flag set. This causes a circular reference, as the instance has a reference back to the object which is its key `_DynamicDataMap` in `KeepAlive` because of this. This is pretty severe, as once an object is used with `DynamicData.For`, it will never be collected by the GC. For Celeste mods, this can cause immense memory leaks, as they usually use `DynamicData.For` quite heavily on e.g. the `Player` object, which in turn has a reference back to the `Level` object, causing it (and the relatively large data structures it holds) to leak as well. The fix is to explicitly specify `keepAlive` to be false for `DynamicData` instances constructed using `DynamicData.For()`.