Fix MonoModRulesModder for nested types
`MonoModRulesModder` is a subtype of `MonoModder` whose purpose is to bypass the relinker for the `MonoModRules` type when it's extracted into its own assembly to be executed as part of the patching process. However, the previous implementation didn't check if the referenced type is a nested type of `MonoModRules`, even though these types are also extracted. Usually, this isn't a problem, as the assembly the type is taken from is itself a mod assembly, for which the relinker uses `FindTypeDeep` to resolve types, which doesn't seem to cause any issues. However, when the `MonoModRules` type is parsed / executed explicitly, usually as part of a runtime relinker, this isn't the case, and as such the output assembly seems to be corrupted in a few subtle ways (presumably related to access modifiers), which can cause accesses to these nested types (including compiler generated types) to fail at runtime with a `TypeAccessException`/ `MemberAccessException`.