Fix failure in .NET 3.5
The IL generated in CreateBclCompileMethodHelper was, in fact, always wrong, it just only mattered on .NET 3.5. When loading a field from a value-type passed in-register, if the field is the only field, the runtime appears to implicitly transform that into a no-op, meaning that the sequence on all new runtimes worked just fine. CLR 2.0, however, requires a property call instead, and that caused us to accidentally dereference the RuntimeMethodHandle instead of just getting its field, leading to a bad pointer.

Fixes #156